refactor: move HUD and input to UI module
